% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanProbaV.R
\name{cleanProbaV}
\alias{cleanProbaV}
\title{Clean Proba-V}
\usage{
cleanProbaV(f_data, QC_val, fill = 255, as.is = FALSE, ...)
}
\arguments{
\item{f_data}{Character. Filename (full path) of a .gif the containing data}

\item{QC_val}{Numeric. Quality control values to \bold{keep} in the data.}

\item{fill}{(vector of) Numeric or NULL. Fill values in the data layer to be filtered out. (important to do before reprojection with resampling method other than Nearest Neighbours)}

\item{as.is}{Logical. As in \code{\link{read.gdal}}.}

\item{...}{Arguments to be passed to \code{\link{writeRaster}}.}
}
\value{
A single rasterLayer, or a stack of the bands for RADIOMETRY input
}
\description{
Uses Proba-V quality layer to perform cleaning over a single data layer.
}
\author{
J Eberenz, adapted from Loic Dutrieux
}
\seealso{
\code{\link{processProbaVbatch}} for batcher
}

