% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype2GDALot.R
\name{datatype2GDALot}
\alias{datatype2GDALot}
\title{Convert raster::dataType to GDAL ot}
\usage{
datatype2GDALot(datatype)
}
\arguments{
\item{datatype}{Character \code{\link{raster::dataType}}}
}
\value{
Character GDAL datatype.
}
\description{
Helper function to use gdal command line tools.
}
\details{
Mapping: In:"LOG1S", "INT1S", "INT1U", "INT2S", "INT2U", "INT4S", "INT4U", "FLT4S", "FLT8S",
Out: "Byte", "Int16", "Byte", "Int16", "UInt16", "Int32", "UInt32", "Float32", "Float64"
}
\examples{
datatype2GDALot("INT2S")
}

