% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extentTogdal.R
\name{extentTogdal}
\alias{extentTogdal}
\title{Convert raster extnet to gdal extent}
\usage{
extentTogdal(x)
}
\arguments{
\item{x}{ratser* or raster::extent object}
}
\value{
numeric vector with corner coordinates (xmin, ymin, max, ymax).
}
\description{
For use with \code{\link{gdalUtils}}
}

