% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHarmMetricsSpatial.R
\name{getHarmMetricsSpatial}
\alias{getHarmMetricsSpatial}
\title{Calculate time series metrics for a tile}
\usage{
getHarmMetricsSpatial(x, n_years = NULL, order = 2, robust = FALSE,
  cf_bands, thresholds = c(-80, Inf, -120, 120), span = 0.3,
  scale_f = NULL, minrows = 1, mc.cores = 1, logfile, ...)
}
\arguments{
\item{x}{RasterBrick or RasterStack Input time series. Layer names must follow proba_v pattern.}

\item{n_years}{Integer. See \code{\link{getHarmMetrics}}}

\item{order}{Integer. See \code{\link{getHarmMetrics}}}

\item{robust}{Logical. See \code{\link{getHarmMetrics}}}

\item{cf_bands}{Integer. Which bands to use for temporal cloud filter (See \code{\link{smoothLoess}})}

\item{thresholds}{threshold Numeric vector of length 2. See \code{\link{smoothLoess}}.}

\item{scale_f}{Numeric of length bands. Sclaing factor per band.}

\item{minrows}{Numeric. Min blcok size fore  \code{\link{mcCalc}}}

\item{mc.cores}{Numeric. Processign cores, see\code{\link{mcCalc}}}

\item{logfile}{Character. See \code{\link{mcCalc}}}

\item{...}{additional arguments to \code{\link{mcCalc}}.}
}
\value{
rasterBrick with metrics as bands.
}
\description{
Spatial version of \code{\link{getHarmMetrics}}, using \code{\link{mcCalc}}.
}

