% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getProbaVinfo.R
\name{getProbaVinfo}
\alias{getProbaVinfo}
\title{Retrieve Proba-v info from filenames}
\usage{
getProbaVinfo(x, pattern, tiles = NULL, ...)
}
\arguments{
\item{x}{Character. Directory of a Proba-V geotiffs or list of filenames.}

\item{...}{Additional arguments to pass to \code{\link{write.csv}}.}

\item{patttern}{Character. As in \code{\link{list.files}}}

\item{tiles.}{Character. Which Proba-V tiles to include. Format: "X00Y00". \code{NULL} for all.}
}
\value{
A data.frame with tile meta data.
}
\description{
Parses through typical filenames and retrieves information on product, tile and acquisition date.
}

