% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProbaVQClist.R
\name{getProbaVQClist}
\alias{getProbaVQClist}
\title{Creates list of ProbaV state mask (SM) codes}
\usage{
getProbaVQClist()
}
\value{
list of clear codes per chanel
}
\description{
Creates list of ProbaV state mask (SM) codes
}
\examples{
QClist <- getProbaVQClist()
}

