% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDistance2Loess.R
\name{mapDistance2Loess}
\alias{mapDistance2Loess}
\title{Spatial calculation of temproal outliers distance}
\usage{
mapDistance2Loess(x, QC_band = NULL, span = 0.3, res_type = c("distance",
  "sd_distance"), mc.cores = 1, ...)
}
\arguments{
\item{x}{RasterBrick. Input ProbaV data, layer names must be Proba-V filenames!
Multiple bands possible, stackign must be chronological.}

\item{QC_band}{Integer. IWhcih band is the SM band?}

\item{span}{Numeric. See \code{\link{smoothLoess}}}

\item{res_type}{Character. "distance" or "sd_distance, see \code{\link{smoothLoess}}.}

\item{mc.cores}{Integer. See \code{\link{mcCalc}}.}

\item{...}{additional agruments to \code{\link{mcCalc}}.}
}
\value{
RasterBrick with distances to loess curve per band and date.
}
\description{
Applies \code{\link{smoothLoess}} spatially using \code{\link{mcCalc}}
to derive the distance of outliers to the \code{\link{loess}} smoothed time
series spatially.
}

