% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcPredictSpatial.R
\name{mcPredictSpatial}
\alias{mcPredictSpatial}
\title{Precdict from raster Brick in parallel}
\usage{
mcPredictSpatial(model, b_metrics, b_clumps = NULL, df_clumps = NULL,
  type = c("response", "probs"), ranger_threads = 1, minrows = 1,
  mc.cores = 1, logfile, ...)
}
\arguments{
\item{model}{Object of class \code{ranger}, see \link{\code{ranger::ranger}}.
Other models are currently not supported!}

\item{b_metrics}{RasterBrick of covariates}

\item{b_clumps}{RasterLayer (optional) with spatial segments}

\item{df_clumps}{data.frame (optional) with additional covarates linked to spatial segments.}

\item{type}{Character What should eb returned, "response" or "probs" (scaled from 0..254).}

\item{ranger_threads}{Integer. Numebr of cores threads for  \link{\code{ranger::predict.ranger}}}

\item{minrows}{Integer. See \code{\link{mcCalc}}.}

\item{mc.cores}{Integer. See \code{\link{mcCalc}}.}

\item{logfile}{Character, see \code{\link{mcCalc}}.}

\item{...}{additional args to \code{\link{mcCalc}}.}
}
\value{
RasterLayer with predictions.
}
\description{
Precdict from raster Brick in parallel
}

