% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeVrtProbaV.R
\name{timeVrtProbaV}
\alias{timeVrtProbaV}
\title{Build GDAL vrt from probav images}
\usage{
timeVrtProbaV(x, pattern, stacked_bands = NULL, vrt_name,
  order_chrono = TRUE, tile = NULL, end_date = NULL,
  return_raster = TRUE)
}
\arguments{
\item{x}{Character. Directory of a Proba-V geotiffs or list of well formated filenames.}

\item{stacked_bands}{Character. Which abnds are stacked (typically RADIOMETRY)}

\item{vrt_name}{Character. Output filename.}

\item{tile}{Character. Which tile to process. Format: "X00Y00".}

\item{end_date}{Date. Last date to process.}

\item{patttern}{Character. As in \code{\link{list.files}}.}

\item{orderChrono}{Logical. Wether to oder the stack chronologically.
Default \code{TRUE}.}

\item{retunr_raster}{Logical. Weather a RasterBrick should eb returned.
Else a info data.frame witht is returned (see \code{\link{getProbaVInfo}})}
}
\value{
a rasterBrick pointing to the vrt or a data.frame with meta data.
}
\description{
Stacks Proba-V layers into a GDAl virtual file,
using \code{\link{gdalbuildvrt}}. GDAL vrt was found to query faster
than \code{raster::stack}.
}
\details{
If stacked bands are present, a nested vrt is build.
}
\examples{


}

