% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHarmMetrics.R
\name{getHarmMetrics}
\alias{getHarmMetrics}
\title{fits robust linear harmonic model to ts and returns the parameters}
\usage{
getHarmMetrics(x, dates = NULL, QC_good = NULL, sig = 0.95,
  n_years = NA, lin_trend = T, order = c(1, 2, 3), robust = F,
  return_model = F, probs = c(0.01, 0.99), ...)
}
\arguments{
\item{x}{Numeric or ts Time series}

\item{dates}{Dates oberservation times if no time series is supplied}

\item{QC_good}{Integer or Logical. If supplied, values of x that are not 1 or \code{TRUE} in QC_good are omitted. Must have same length as x and dd.}

\item{n_years}{Integer. If not \code{NULL}, additional coeff. for n intra-yearly variation. NOTE: Not working yet}

\item{lin_trend}{Logical. Should a linear trend be modelled?}

\item{order}{Numeric. First or second order harmonics}

\item{robust}{Logical Use robust regression from \code{\link{robustbase:lmrob}}? NOTE: Not working yet}

\item{probs}{Numeric. Vector of two probabilities to compute quantiles.}

\item{...}{additional args. non implemented.}
}
\value{
Numeric vector of the models coefficients and percentiles (length depending on arguments)
}
\description{
Processes Proba-V data for subsequentent use in time-series analysis. Performs Proba-V cleaning and operations with parallel support.
}
\seealso{
\code{\link{removeDips}}, and \code{\link{robustbase::lmrob}}
}

