% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDistance2Loess.R
\name{mapDistance2Loess}
\alias{mapDistance2Loess}
\title{Title}
\usage{
mapDistance2Loess(x, QC_band = NULL, span = 0.3, res_type = c("distance",
  "sd_distance"), mc.cores = 1, ...)
}
\arguments{
\item{x}{RasterBrick.}

\item{QC_band}{}

\item{span}{}

\item{mc.cores}{}

\item{...}{}

\item{yday}{Numeric. See getHarmMetrics}

\item{n_years}{}

\item{order}{}

\item{robust}{}

\item{threshold}{}
}
\description{
Title
}
\examples{


}

