% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/probaVTileFromCoords.R
\name{probaVTileFromCoords}
\alias{probaVTileFromCoords}
\title{Proba-V tile from coordinates}
\usage{
probaVTileFromCoords(x, y)
}
\arguments{
\item{x}{Numeric. Longitude}

\item{y}{Numeric. Lattitude}
}
\value{
A character vector with the Proba-V tiles, format \code{"X00Y00"}
}
\description{
Get Proba-V tile from coordinates
}
\examples{
probaVTileFromCoords(10, -5)
}
\author{
J Eberenz
}

