% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeVrtProbaV.R
\name{timeVrtProbaV}
\alias{timeVrtProbaV}
\title{Build GDAL vrt from probav images}
\usage{
timeVrtProbaV(x, pattern, stacked_bands = NULL, vrt_name,
  order_chrono = TRUE, tile = NULL, start_date = NULL, end_date = NULL,
  return_raster = TRUE, ...)
}
\arguments{
\item{x}{Character. Directory of a Proba-V geotiffs or list of well formated filenames.}

\item{pattern}{Character. As in \code{\link{list.files}}.}

\item{stacked_bands}{Character. Which bands are stacked (typically RADIOMETRY)}

\item{vrt_name}{Character. Output filename.}

\item{tile}{Character. Which tile to process. Format: "X00Y00".}

\item{start_date}{Date. First date to process.}

\item{end_date}{Date. Last date to process.}

\item{return_raster}{Logical. Wether a RasterBrick should be returned.
Else an info data.frame is returned (see \code{\link{getProbaVInfo}})}

\item{order}{Chrono Logical. Wether to order the stack chronologically.
Default \code{TRUE}.}
}
\value{
a rasterBrick pointing to the vrt or a data.frame with meta data.
}
\description{
Stacks Proba-V layers into a GDAl virtual file,
using \code{\link{gdalbuildvrt}}. GDAL vrt was found to query faster
than \code{raster::stack}.
}
\details{
If stacked bands are present, a nested vrt is build.
}
\examples{


}
\author{
J Eberenz
}

